Make sure thumbnails work when images are on S3
Copied from q-m/wp-json-api@ff9ffe1 -- I would have preferred to create a PR with his commit but he didn't for this repo so I had to re-create the change. This commit is necessary for us to implement an api while using an S3 plugin that pushes all images to S3.  The actual files don't exist locally so without this change, we get an empty array and the only available `url` is one that's local and is not valid.